pyproject.toml
[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "Prototype: autonomous data analysis agent (LangGraph) + Streamlit UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "numpy>=1.26.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "scikit-learn>=1.4.0",
  "tabulate>=0.9.0",
  "streamlit>=1.36.0",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"


